1. Added Exception Mapper

+++ customer-handler-service-controller/src/main/java/com/customer/handling/service/controller/AddressController.java
@@ -1,6 +1,7 @@
 package com.customer.handling.service.controller;
 
 import com.customer.handling.service.api.AddressData;
+import com.customer.handling.service.exception.RequestNotValidException;
 import com.customer.handling.service.mapping.ApiControllerModelMapper;
 import com.customer.handling.service.service.AddressService;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -10,6 +11,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.validation.constraints.NotNull;
 
 @Slf4j
 @RestController
@@ -25,7 +28,8 @@ public class AddressController {
     @Tag(name = "method.POST", description =  "post Data body to My SQL DB with help Springdoc OpenApi:-P")
     @PostMapping(value = "/create", produces = MediaType.APPLICATION_JSON_VALUE, consumes = "application/json")
     public ResponseEntity<AddressData> createAddress(@RequestBody AddressData addressData) {
-        AddressData addressCreateData = apiControllerModelMapper.map(addressService.createAddress(apiControllerModelMapper.map(addressData)));
+        AddressData addressCreateData = apiControllerModelMapper
+                .map(addressService.createAddress(apiControllerModelMapper.map(addressData)));
         log.info("create address with body: " + addressData);
         return ResponseEntity.ok(addressCreateData);
     }
@@ -33,7 +37,10 @@ public class AddressController {
     @ResponseBody
     @Tag(name = "method.GET")
     @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<AddressData> getAddress(@PathVariable("id") Integer id) {
+    public ResponseEntity<AddressData> getAddress(@PathVariable("id") @NotNull Integer id) {
+        if (id <= 0) {
+            throw new RequestNotValidException("Id must not be negative or Zero");
+        }
         AddressData addressGetData = apiControllerModelMapper.map(addressService.getAddress(id));
         log.info("get request address body with DbId: " + addressGetData);
         return ResponseEntity.ok(addressGetData);
@@ -41,7 +48,10 @@ public class AddressController {
 
     @Tag(name = "method.DELETE", description =  "delete Data body from My SQL DB")
     @DeleteMapping(value = "/{id}")
-    public void deleteAddress(@PathVariable("id") Integer id) {
+    public void deleteAddress(@PathVariable("id") @NotNull Integer id) {
+        if (id <= 0) {
+            throw new RequestNotValidException("Id must not be negative or Zero");
+        }
         addressService.deleteAddress(id);
         log.info("delete request address body with DbId: " + id);
     }
@@ -49,8 +59,12 @@ public class AddressController {
     @ResponseBody
     @Tag(name = "method.PUT")
     @PutMapping(value = "/update", consumes = "application/json", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<AddressData> updateAddress(@RequestBody AddressData addressData){
-        AddressData addressDataUpdated = apiControllerModelMapper.map(addressService.updateAddress(apiControllerModelMapper.map(addressData)));
+    public ResponseEntity<AddressData> updateAddress(@RequestBody AddressData addressData) {
+        if (addressData.dbId() <= 0) {
+            throw new RequestNotValidException("Id must not be negative or Zero");
+        }
+        AddressData addressDataUpdated = apiControllerModelMapper
+                .map(addressService.updateAddress(apiControllerModelMapper.map(addressData)));
         log.info("update request address body with DbId: " + addressData);
         return ResponseEntity.ok(addressDataUpdated);
     }
